refactor and move to right package
