--- app/src/main/java/com/example/financewallet/HomeFragment.kt
@@ -1,57 +1,0 @@
-package com.example.financewallet
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.financewallet.databinding.FragmentHomeBinding
-
-class HomeFragment : Fragment() {
-
-    private lateinit var binding: FragmentHomeBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment using ViewBinding
-        binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.portfolioListBtn.setOnClickListener {
-            loadFragment(PortfolioListFragment())
-        }
-        binding.settingsBtn.setOnClickListener {
-            loadFragment(SettingsFragment())
-        }
-        binding.assetListBtn.setOnClickListener {
-            loadFragment(AssetListFragment())
-        }
-    }
-
-    private fun loadFragment(fragment: Fragment) {
-        parentFragmentManager.beginTransaction()
-            .replace(R.id.fragment_container, fragment)
-            .addToBackStack(null)
-            .commit()
-    }
-
-    companion object {
-        private const val ARG_PARAM1 = "param1"
-        private const val ARG_PARAM2 = "param2"
-
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            HomeFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
-    }
-}
