+++ app/src/main/java/com/example/financewallet/presentation/assetList/AssetListFragment.kt
@@ -1,0 +1,27 @@
+package com.example.financewallet.presentation.assetList
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.financewallet.databinding.FragmentAssetListBinding
+
+class AssetListFragment : Fragment() {
+
+    private var _binding: FragmentAssetListBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentAssetListBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
