+++ app/src/main/res/layout/fragment_asset_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AssetListFragment">
+    tools:context=".presentation.assetList.AssetListFragment">
 
     <TextView
         android:layout_width="match_parent"
