+++ app/src/main/res/layout/fragment_portfolio_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PortfolioListFragment">
+    tools:context=".presentation.portfolio.PortfolioListFragment">
 
     <TextView
         android:layout_width="match_parent"
